I removed all the comments that say "Known magic number" from the KK code.
Magic number -- definition:  a value hard-wired into the computer code that has
                             been arbitrarily chosen for a parameter.  The values
                             are not the result of a mathematical calculations, such
                             as differentiation or integration, but rather are
                             "seems-like-it-works" numbers.

I can only assume that the "magic numbers" that were being referred to in those
comments were the constant coefficients in the equations.  However, these
constant coefficients are NOT "magic numbers".  Rather, the are the constant
coefficients that result from integration of a function.  Magic numbers are changable
numbers (which is why it is a good idea not to hard-wire them), whereas these
numbers won't change.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@5459 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 1e61139a2d314abbb7f34f61b636e618a2a5e79b