Multiple changes for ticket #485.
Because the LES data for the cloud_feedback cases only goes out to 10 days for two of the cases, we have shortened the runtime of all the cloud_feedback cases to 10 days.

Adding the LES lines for the cloud_feedback cases also required some changes to the inner workings of plotgen.  The way altitude is determined has been changed from looking specifically at the second field in the netcdf file to finding the 'z' variable and reading that variable specifically.  This change required another if case for twp_ice because the variable name in this file is 'altitude'.

Secondly, the cloud_feedback LES output has only two dimensions(time and z), whereas the other LES data has 4 dimensions (x, y, z and time).  I added an if statement to the averaging portion of plotgen to handle the 2 dimensional cases since plotgen was erroring out on the 'squeeze' function in 2 dimensional cases because it was trying to access 4 dimensions.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@5664 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 656af27653bcbf3a99a60d5b47a944bfc6e37b5c