I added a new subroutine that handles the setup for upscaled KK.  This includes the unpacking and setting of means, standard deviations, and correlations. This also includes the normalization of appropriate means, standard deviations, and correlations.
This is part of a new setup for KK, which will allow a user to bypass the
calls to KK mean tendencies and simply call covariances.  These variables are
needed for both upscaled means and upscaled covariances.

I ran tests cases for DYCOMS-II RF02 DO and DS, as well as RICO.  The GrADS
result files matched to binary exactness.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@5680 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 4eaa80e0ad203e0fe9a1d2b07cd55078e24f6a1d