I added some new files: 1) abs_val_signed_sort.m sorts an input vector of values by absolute magnitude, keeping the sign of each value the same. 2) sort_sum.m sums the values sorted by abs_val_signed_sort.m.  Summing from smallest magnitude to largest usually increases the accuracy of numerical addition 3) sort_group_sum sums the values sorted by abs_val_signed_sort.m, but does so in such as way as to take partial sums of different order-of-magnitude groups, which greatly increases the accuracy of numerical addition.
These methods are more accurate than MATLAB's sum, which simply sums values in
the order they are found in the input vector.

This new functions are going to be used in my microphysics function integration
files, in order to increase the accuracy of the numerical integration portion of
those functions.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@5850 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 9e0d636e2ee972983a8d40ebdba16fc8b017d69c