I noticed a bizarre problem starting popping up recently.  The nightly tests have been running successfully for every compiler except for the Sun compiler. Using the Sun compiler, the code successfully compiles, but then fails to run for four cases:  dycoms2_rf02_do, dycoms2_rf02_ds, rico, and astex_a209.  The traceback from the failure shows that the problem has to do with writing stats to GrADS.  The one thing those four cases have in common is KK microphysics. The problem didn't start popping up until I added a bunch of new stats recently to KK micro.  The problem had to do with some of the new stats -- specifically mu_rr_n, mu_Nr_n, and mu_Nc_n.
Let's use mu_rr_n as the example here.  When there is sufficient level mean
rain water mixing ratio (rrainm > rr_tol), there will be a sufficient mean
in-precip rain water mixing ratio (mu_rr).  The mean (in-precip) of ln r_r
is mu_rr_n.  This number should be negative, but not enough to cause a
problem when there is sufficient rain at a level.  However, when there is
no rain at a level (or less than tolerance), mu_rr is 0 and mu_rr_n should be
negative infinity.  When this scenario occurs in the code, mu_rr_n is simply
set to -huge( mu_rr_n ), the largest magnitude negative number that can be
represented by the given variable kind.  This is alright because mu_rr_n is not
used again in the code -- only output to stats.

The problem occurs when the CLUBB core_rknd is double precision by default.
When there isn't any rain at a given grid level, mu_rr_n is set to -huge and
has a value on the order of -10^308.  When this is output in stats, GrADS
writes data in only single precision (where the largest value is on the order
of -10^38).  When the Sun compiler is used (and only this compiler in debug
compilation setting), a runtime floating point exception is caused.

The problem is rectified by setting mu_rr_n, mu_Nr_n, or mu_Nc_n to
-huge( 0.0 ), which sets them to the largest magnitude negative number that can
be represented by single precision, when the rrainm, Nrm, or Ncm is less than
its respective tolerance amount.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6043 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: c2239591bf11f82529ce9231479f97410b101f45