The terms -d<V_rr'rr'>/dz and -d<V_Nr'Nr'>/dz are going to be added to the predictive equations for <r_r> and <N_r>, respectively.  The covariance between V_rr and r_r is a constant times the covariance between r_r and KK rain drop mean volume radius (and likewise for N_r).
Here, the code to calculate and keep stats for the covariances
<r_r'R_vr'> and <N_r'R_vr'> (where R_vr is KK rain drop mean volume radius)
is being added to the SVN repository.  Additionally, covariances
<V_rr'r_r'> and <V_Nr'N_r'> are calculated (on thermodynamic levels) based
on the covariances involving mean volume radius.  These covariances
involving sedimentation velocities will need to be interpolated to momentum
levels so their derivatives can be taken over the central thermodynamic levels.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6064 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 2fcd22539e587397081d061babc840cb78ff172a