The terms -d<V_rr'rr'>/dz and -d<V_Nr'Nr'>/dz are now added to the predictive equations for <r_r> and <N_r>, respectively.  The covariance between V_rr and r_r is a constant times the covariance between r_r and KK rain drop mean volume radius (and likewise for N_r).
The remainder of the code to add these terms is complete!  The hydrometeor
covariance terms are passed into microphys_solve and added to the RHS of the
predictive equation.  The terms are also added to the statistical budgets for
the appropriate predictive equations.  Both the <r_r> and <N_r> budgets still
balance.  The terms are also passed into the subroutine adj_microphys_tndcy,
which accounts for the over-evaporation of rain and provides a microphysical
source adjustment term to compensate for over-evaporation.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6070 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 7e061521bb0c27330eaee4b486dc0bd2733e2ff8