The correlations used within KK microphysics may differ sometimes from the correlations prescribed/diagnosed in the correlation array.  This is because:
1)  KK uses correlations for each component.  Right now, KK selects either the
within-cloud or outside-cloud (below-cloud) value for a correlation for EACH COMPONENT
based on whether COMPONENT cloud water mixing ratio is at or above tolerance level,
not overall mean cloud water mixing as is used in the correlation array.

2)  When rain water mixing ratio (and/or rain drop concentration) in a component is less
than tolerance amount, it is considered to have a value of 0.  In this scenario, the
values of the correlations involving rain water mixing ratio (and/or rain drop
concentration) are set to 0, just to be on the safe side.  This makes the correlation
differ from the correlation value found in the correlation array.

3)  I am experimenting with a code that will set the value of a correlation in a component
to a value that is the within-cloud value, outside-cloud value, or an interpolated value
in-between those two based on component cloud fraction.  This will also make the value of
a correlation differ from the value found in the correlation array.

I is beneficial to have the values of the correlations that are actually used in the code
output to statistics.  For the above reasons, I have restored the stats corr_srr_1,
corr_srr_2, corr_sNr_1, corr_sNr_2, corr_sNc_1, corr_sNc_2, corr_rrNr_1, and corr_rrNr_2
to subroutine KK_stat_output (they recently had been moved out of there).  The stats that
record the correlation array values for the above correlations in subroutine corr_stat_output
are now simply called corr_srr, corr_sNr, corr_sNc, and corr_srr.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6170 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 454976e5290360bbe39f0937fb1d3fc02015586b