The covariances of KK rain drop mean volume radius and r_r or N_r are always realizable, as the are calculated by the PDF, as is the variance of KK rain drop mean volume radius.  This was expected.
Since <r_r'^2>, <N_r'^2>, and <R_vr'^2> are not necessary to check the limits of
the covariances <R_vr'r_r'> and <R_vr'N_r'>, they have been moved out of the
subroutine sed_vel_covars and into KK_upscaled_stats.  They are still being kept
as statistical output.  The position of KK_upscaled_stats was moved to the end
of the microphysics tendency loop.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6176 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: fc9bb801ae0fe1d6423b18cecf5458d59f64a2f1