I have added a new option to the KK code.  This option calculates the covariances involving sedimentation velocity ( < V_rr'r_r' > and < V_Nr'N_r' > ) as an equation of the form: < V_hm'h_m' > = coef_A * < h_m > + term_B. This allows for the turbulent sedimentation of < h_m > to be solved in a semi-implicit manner, increasing the numerical stability of the solution. The covariance < V_hm'h_m' > is written in this way by simply rearranging the original equation.
In order to compare the results using the new form to the results using the
original form, the if...else statement that selects the new or original form
was commented out (so both forms were calculated), and print statements were
placed in the code.  The new form was calculated using:
< V_hm'h_m' > = coef_A * < h_m >(t) + term_B;
with < h_m > being from the current timestep (completely explicit), making it
easy to compare the new and original forms of the covariance.

The results consistently matched with percent differences on the order of
1.0e-13% to 1.0e-14% !!!!!



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6186 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 361d786a3b877e7183f6eecd99ec89ce93b8a776