In the CLUBB hydrometeor mean field equations, the turbulent advection term, < w'h_m' > is handled according to a down-gradient assumption.  This allows the turbulent advection term to be treated as an eddy-diffusion term in the predictive equation set.
I wanted to produce the values of wprrp, wpNrp, etc. according to:
wprrp = -K_r drrm/dz.

In order to do so accurately, I handled the turbulent advection of the mean
hydrometeor field and the calculation of covariance wphmp using a
Crank-Nicholson time-stepping scheme, as was previously done with um/vm and
upwp/vpwp.

Additionally, I increased the outside-cloud correlations for s,rr and s,Nr.
I increased the in-cloud ratio of rrp2 to rrm**2 to 1.0 (and likewise for Nr).
This is an increase over the DYCOMS-II RF02 values that were previously being
used for RICO.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6208 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: a159ae4af5be9daf689394ac5c80f7f39a241bdc