I was planning to remove wpthlp_mc_tndcy, wprtp_mc_tndcy, rtp2_mc_tndcy, thlp2_mc_tndcy, and rtpthlp_mc_tndcy from the Morrison and local KK codes, since only upscaled KK can output these terms.  They were previously added as output from Morrison micro. because KK upscaled and KK local were intertwined, and any code that could be used with Latin Hypercube needed to have uniform subroutine call lists.  However, when I was attempting to remove these terms from the Morrison and local KK (now separated from upscaled KK), a found some code in Morrison that outputs rtp2_mc_tndcy and thlp2_mc_tndcy (simple attempts to reproduce that grandeur of upscaled KK covariances).  This could cause a conflict.
When Latin Hypercube is used with Morrison microphysics (or any local
microphysics scheme), the covariances of each of w, r_t, and theta_l
with each of evaporation rate, autoconversion rate, and accretion rate
need to be produced using sample points from the PDF.  Then, the
values of wpthlp_mc_tndcy, wprtp_mc_tndcy, rtp2_mc_tndcy, thlp2_mc_tndcy,
and rtpthlp_mc_tndcy are calculated from the aforementioned covariances.
This is not compatible with the code found with the l_morr_xp2_mc_tndcy
flag turned on.  I entered a safeguard in the code so that interactive
Latin Hypercube cannot be used along with l_morr_xp2_mc_tndcy.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6216 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 03f2896d7bf66ab81296e82dcc61cd9b776f3d53