I renamed subroutine KK_upscaled_setup as subroutine normalize_PDF_params.
Correlations involving w and a hydrometeor or t and a hydrometeor are now
normalized in this routine, along with the other correlations (as well as means
and variances) requiring normalization, so that all normalization happens in one
subroutine.  Previously, the correlations involving w and a hydrometeor or
t and a hydrometeor were normalized in the KK_upscaled_covar_driver, since those
correlations were only needed for that code.




git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6263 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 4e809c0192a798be7328e4744d5de5eab9d9f373