Corrected a small error in plotgen. For timeseries plots, the values on the time axis and the values plotted are potentially inconsistent, particularly when the user specifies start and end times that are not multiples of the time step. In the code, the time indexes were calculated as: t_start_index = ceil(startTime / time_step_length); t_end_index = round(endTime / time_step_length);
but the time array for the x-axis of time timeseries plots was calculated as:
times = startTime:time_step_length:endTime

This is inconsistent. I changed the x-axis time array to be calculated as:
times = (t_start_index:t_end_index) * time_step_length;

The use of "round" in t_end_index is also inconsistent with the way that the times are pulled from the file. Therefore, I changed t_end_index to use "ceil" instead.

This should fix the three plots of ARM_CC that do not show up in the nightly tests as mentioned in cpt:comment:14:ticket:161 (they will show up as 0's right now because the variable is not found.)


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6280 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: c7147904ce7a1bfb7e3e529b1e14b87b993cea5c