I have modified the `l_use_modified_corr` code to make Cholesky decomposition successful.
I use the method in SILHS to estimate the correlation between `t` and `rr` and between `t` and `Nr`.

I also limit the correlation between `s` and `t`, because a correlation matrix with a perfect correlation between `s` and `t` cannot be decomposed.

Output is bin-for-diff identical when `l_use_modified_corr` is not used. Also, the assertion check for SILHS cloud weighted sampling no longer complains.

See ticket:514.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6469 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: b6fd28a86ededd8aadde70d14701af5f0f6fa2d6