I am getting rid of the special lower boundary condition for hydrometeors (at the k = 1 thermodynamic level, which is below the model lower boundary (or below the surface)).  When both l_sed and l_upwind_diff_sed were set to .true., sedimentation of hydrometeors took place at the k = 1 thermodynamic level, which is odd because the level is below the lower boundary and no other microphysical process rates are calculated there.  Anyway, this would cause hydrometeors to "pile up" at the k = 1 thermodynamic level in some cases, creating an odd spike on the profiles of mean rain water mixing ratio and mean rain drop concentration.
I got rid of the special condition and simply enforced the code that was
being used at the lower boundary for centered differencing -- the value of
the hydrometeor at thermodynamic level k = 1 is set equal to the value of
the hydrometeor at thermodynamic level k = 2, which is above the model lower
boundary (above the ground).

Obligatory Larson Group Internal Trac System Reference: clubb:ticket:635.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6629 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: f98c01d78e63d9f5f3392fec06505dbc9e4b8236