For clubb:ticket:603, I added a Ngraupelm plot to the deep convective case files.
It is worth noting that the conversion between SAM and CLUBB number concentrations is different
in twp_ice than in the other cases.  I assume the twp_ice case was run with different microphysics,
but if we rerun the twp_ice case with SAM using Morrison, we will have to change the factor to match
the other cases.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6657 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 83659cf554d04f5062bd20656d0a27cd5857653f