I am making a significant change to CLUBB.  I am splitting off Ncm from the hydromet array.  As part of this change, the iiNcm hydromet array index has been eliminated.  Pieces of code that were previously controlled by "if ( iiNcm > 0 )"-type statements are now controlled by "if ( l_predictnc )". There is now a new block of code to setup the LHS, RHS, and solve for Ncm when l_predictnc is enabled.  This is done after the hydrometeor solve loop in subroutine advance_microphys.
The GrADS output results for this revision are an exact match to the GrADS
output results from the previous revision for all cases with one notable
exception -- mpace_b_LH.

The results of mpace_b_LH look significantly different (and much worse)
for Ncm and Nc_in_cloud.  This is somewhat complicated to explain.  Along the
way, I discovered a bug in Latin Hypercube -- Morrison code.  This had to do
with one small line of poor coding within our Morrison microphysics driver.
When I corrected this bug, the results of mpace_b_LH case got much worse.  I
think this is due to a second bug in the initialization of Ncm in CLUBB.  I
have not yet tried to correct this.

Obligatory Larson Group Internal Trac System Reference: clubb:ticket:658.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6771 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 1da9599c4fa0100878155c78714e35b41a484e7c