I am doing two things with this commit:
1)  I am updating cloud_sed_module.F90 to much more modern coding style (this
was written about 8 or 9 years ago).  This GrADS file results from this portion
of this revision are an exact match to the GrADS file results from the previous
revision.

2)  I am updating the call to cloud_drop_sed so that it is called with
Nc_in_cloud * cloud_frac rather than Ncm.  This keeps the mean cloud droplet
concentration passed into cloud_drop_sed "up-to-date" or consistent with rcm
and cloud fraction.  This should fix the budget issues for the DYCOMS-II RF02
DS case.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6799 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 911f00b537925ad3f13fcc427f04361de365da8c