I am organizing the CLUBB microphysics code.  I am moving the code that advances a mean microphysics field one model timestep (advance_microphys, advance_hydrometeor, advance_Ncm, microphys_solve, microphys_lhs, microphys_rhs, and the sedimentation and turbulent sedimentation functions) to its own module and file, advance_microphys_module. This code is modeled after the "advance_" files and subroutines found in CLUBB_core.
The remaining code in microphys_driver.F90 is subroutine microphys_schemes, which calls
a microphysics scheme and outputs tendencies.  The call to microphys_schemes is made
from subroutine run_clubb (clubb_driver.F90), which is followed by the call to
advance_microphys.

The results of GrADS file output for this revision is an exact match to the results of
GrADS file output for the previous revision for all cases.

Obligatory Larson Group Internal Trac System Reference: clubb:ticket:671.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6808 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 0f7ac81c46f81f3b59ffa67f20fb9a83d4f4c34f