I am committing a slight improvement to subroutine init_microphys.
The setting for hydrometeor array hydromet_name, l_mix_rat_hm, and
l_frozen_hm are now set after the "select case" block for micro_scheme,
where hydromet_tol is set.  Unlike the ii indices (iiNrm,etc.) and
the l_hydromet_sed array, the values for these arrays do not depend on
the microphysics scheme, but only on whether a hydrometeor is predicted
(which we can tell by iiNrm > 0, for example) and present in the
hydromet family of arrays.

The GrADS output file results from this revision are an exact match to
the GRADS output file results from the previous revision.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6812 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: e72afda338366844d1ed4977528d01451ac0bf5c