I am taking care of our budget balancing issue for ice crystal concentration.
The problem arises because the tolerances used in checkBudget.py are way too
small for the hydrometeor concentrations.  Recently, the small allowable positive
value of a hydrometeor concentration was changed in CLUBB to be based on the
smallest allowable positive mixing ratio and the a maximum allowable mean volume
radius for the hydrometeor.  For ice, this means that the smallest allowable
positive concentration is on the order to 1.0e-2 num/kg.  Additionally, typical
values of ice concentrations are much larger than this minimum value.

However, rather than use an acceptable percentage difference (which would be better),
the checkBudget.py code uses an acceptable absolute difference.  The absolute
difference that was being used for frozen hydrometeor concentrations was 1.0e-7 num/kg.
This results in a warning message being printed even when Nim_bt and the sum of the
budget terms are off way down in the nth decimal place.  For now, I have increased
the tolerance used for ice.  However, I think a better solution using a percent difference
should be adopted.  For example:

percent difference
= Nim_bt - sum_of_budget_terms ) / max( sum_of_budget_terms, eps ) * 100

when percent_difference > percent_diff_tol, flag it!

Obligatory Larson Group Internal Trac System Reference: clubb:ticket:647.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6844 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 0750abc989654742394098e67df65fa81504aa04