I recently ran across a problem in CLUBB.  When looping over all the hydrometeors in hydromet_dim and mapping the hydromet index to the PDF index, I ran across an array out of bounds error.  In a nutshell, it turned out that the hydrometeor set of indices (iirrainm, iiNrm, etc.) was not always entirely subset of the PDF set of indices (iiPDF_rrain, iiPDF_Nr, etc.).  To make a point, each index in the hydrometeor set of indices should also correspond to an index in the PDF set of indices.  In other words, the hydrometeor set of indices should be a proper subset of the PDF set of indices. When there are indices found in the hydrometeor set that are not found in the PDF set, this causes problems.
Anyway, this problem was occuring when Morrison microphysics was run with l_graupel
turned off.  The hydrometeor indices were not set up properly for this specific
scenario, whereas the PDF set of indices were set up properly to handle this.

To rectify the situation, a little refactoring was required.  One of the places that
required refactoring was morrison_micro_driver_module.F90.  Since this is
important code, I decided to commit this file by itself so I could more easily double
check that I am not changing any results.

The GrADS output file results from this revision are an exact match to the GrADS
output file results from the previous revision for all cases.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6892 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 09d37b00d6901fdf16bc488d4e6e80cecd726a26