This had to be the simplist fix to a troublesome problem of not reading in the prescribed correlations involving Ng.  Basically, all the prescribed correlation files (case_corr_array_cloud.in, etc.) use the header "Ng", while the code in CLUBB that matches based on that name was trying to match to "Ngraupel".
The solution was to either truncate "Ngraupel" to "Ng", or to increase all the
"Ng"s in the corr_array.in files to "Ngraupel"s.  The corr_array.in files are
already right up against the 100 character limit, so we need to delete 6
characters in corr_matrix_module.F90 that make all the difference for proper
Ng correlations being read into the code.

Obligatory Larson Group Internal Trac System Reference: clubb:ticket:687.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6898 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: da3c7c430ebc00fef4bbf5b0b44025fc50fb710f