I am committing a change to the Morrison microphysics hydrometeor indices.
Not too long ago, I ran across a problem in CLUBB. When looping over all the
hydrometeors in hydromet_dim and mapping the hydromet index to the PDF index, I
ran across an array out of bounds error. In a nutshell, it turned out that the
hydrometeor set of indices (iirrainm, iiNrm, etc.) were not entirely a subset of
the PDF set of indices (iiPDF_rrain, iiPDF_Nr, etc.). To make a point, each
index in the hydrometeor set of indices should also correspond to an index in
the PDF set of indices. In other words, the hydrometeor set of indices should be
a proper subset of the PDF set of indices.  When there are indices found in the
hydrometeor set that are not found in the PDF set, this causes problems.
Anyway, this problem was occuring when Morrison microphysics was run with
l_graupel turned off. The hydrometeor indices were not set up properly for this
specific scenario, whereas the PDF set of indices were set up properly to handle
this.

I this revision, the hydrometeor set of indices are setup to take into account
the l_ice_microphys and l_graupel flags for Morrison microphysics.

The results change for some cases, mainly due to SILHS sample points (random
numbers are associated with different hydrometeors in this revision than they
were in the previous revision due to the reordering).  There are also some small
differences due to numerical round off error when hydrometeor mixing ratios are
added in a different order.  For some unstable cases, such as MC3E, this results
in differences that are somewhat magnified.




git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6907 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: be79fd908904270b2b1920b277fe7e927739b4d5