For clubb:ticket:607. I found three subroutines in CLUBB in which the variable, err_code, occurred in an incorrect location. Err_code has been moved in the argument list of these subroutines and code which calls the subroutines has been updated. Output is bindiff identical.
It would be possible to add a CLUBBStandardsCheck script which would alert us if any variables in an argument are in an unexpected order anywhere in CLUBB. Although this script would not have to run every time CLUBB is compiled, it could be placed in the utilities/check_scripts folder. Would this be worth while?

git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6915 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 1ec010d7988fc9f2da2a1ff4488c62380112511c