I added a namelist and a run script for G_unit_tests.
The namelist will allow people to more easily select which unit tests to run.
Currently, the code needs to be recompiled to turn an individual tests on or off.

The runscript is used to make running G_unit_tests easier.  It copies the namelist
to the local run_scripts directory, runs G_unit_tests, and then removes the local
copy of the namelist.  Most of the other CLUBB executable programs also have
run_scripts.  This might make G_unit_tests easier to find for an unfamiliar user.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6934 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 833576ba518ce41f7057deb973b3889314e03723