We have a check that ensures that the correlation between `s` and `Ncn` is zero when generating a correlation matrix.
If not, we warn the user that the SILHS solution might not converge to the "analytic" solution (upscaled KK, I presume).

I made this a debug level 2 check. I also got rid of the lines that ask the user whether they want to continue, as such prompting is irregular (in that we don't do this anywhere else in the code) and causes automated scripts to fail.

See cam:ticket:8, particularly [cam:comment:26:ticket:8 comment:26].


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6997 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: cf1c321bc463d34c16612c1387b1e9c9d6aed0e7