For clubb:ticket:680. Added comments to all of the variables I could (excluding the variables only in clubb standalone) in the CLUBB API. Could someone take a look at the comment for l_silhs_KK_convergence_adj_mean in the CLUBB API  and be sure that what I wrote is accurate as well as come up with a one sentence description for the function, latin_hypercube_2D_output, in the SILHS API?
I've found two good reads on API's, which I'll link to below. Although they are both from an OOP and initial code design standpoint, I think the principles remain similar.

The first is a talk given by the principle software engineer at Google, Joshua Bloch. Slides 28 (Don't Make the Client Do Anything the Module Could Do) and 35 (Avoid Long Parameter Lists) seem to be especially important to us. The slides can be found at:
http://lcsd05.cs.tamu.edu/slides/keynote.pdf

The second is "The Little Manual of API Design" by Jasmin Blanchette. If I were to only read one chapter, I'd read chapter 4. What I especially like about that chapter is that it tends to provide a counter-opinion to much of what is said in Bloch's slides. No one seems to agree with anyone else when it comes to API design. The pdf of the book can be found at:
http://www4.in.tum.de/~blanchet/api-design.pdf

At this point, I am pretty much done working on the API's and I'd like to start integrating them with CLUBB_standalone.

git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7041 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: f81fe71138a26284279063bf6df941b1bcff65b7