I refactored the way that SILHS performs some of its assertion checks.
Perhaps the coolest part of this change is that the check for equal number of sample points in/out_of cloud in normal space has been moved to `lh_subcolumn_generator`. Also, it now only runs when `clubb_at_least_debug_level( 2 )`.

Output is bit-for-bit identical for (at least) `arm_97` and `rico_lh`. Also, I manually tested the code with purposely mutated variables and the assertions all triggered successfully.

Some subset of this commit is in partial or complete fufillment of the requirements outlined in clubb:ticket:716.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7097 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: fb5fc842d4b4917ab4ec272de452b0d3081caaf3