For clubb:ticket:720, I am imposing a tolerance on `stdev_chi_i`, assuming the distribution of `chi` to be a delta if `stdev_chi_i < chi_tol`.
This fixes an inconsistency that was present in the code between `cloud_frac_i` and the distribution of `chi`. The inconsistency resulted in SILHS sampling the PDF improperly in some remote corner cases where the standard deviation of `chi` was small.

Output is bin-diff identical for most cases, but not all of them. ARM-97 and TWP-ICE are notable exceptions.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7143 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 3d3ab3494083e54a851d617a9dadfadbd89f4c46