For clubb:ticket:736, I fixed a bug I injected into the SILHS code.
In r7265, the logic to determine whether to do importance sampling in cloud was moved into `cloud_weighted_sampling_driver`. `X_u_chi` appears as `intent(out)` of this `subroutine`, therefore it must be set. However, if it is determined that `cloud_frac` is not in the proper range, it will not be set, and the `intent(out)` variable will be left uninitialized.

To fix this problem, I made `X_u_chi` and `X_u_dp1` `intent(inout)` of `cloud_weighted_sampling_driver`.

This resolves the failure of the Oracle Bitten test last night. This bug did not change output on "normal" compilers that don't automatically clobber `intent(out)` variables.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7267 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 16aa4f0759a819b89eaab7807be268e6406f0c83