I added statistical output for wp2hmp, rtphmp, and thlphmp (for all hydrometeors, hm, found in the microphysics scheme).
The cases are not changed, but additional stats are produced.

Continuing with our series ...

The Kurt Kotenberg Memorial Trac System Reference: clubb:ticket:639,
but it also applies to clubb:ticket:730.

Kurt Kotenburg was filled with humor and a passion for pork chop sandwiches!
He graduated from UWM with a Master's Degree and became an on-air T.V.
meteorologist at WEAU T.V. in Eau Claire.  He turns up on a quick Google
search as a National Weather Service meteorolgist in Des Moines.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7292 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 52747ebe269a3cfe0c5276cceb554636542fa979